Move build_rebuild_siblings_jql to appropriate place
Importing it from jira tools caused agent docker image to
require flexmock.
